Shifted tests and made key variables accessible from outside
