+++ tests/test_data.py
@@ -1,13 +1,57 @@
 import os
+from contextlib import contextmanager
 
 import numpy
 import pytest
+from RIAssigner.data import Data
 from RIAssigner.utils import get_extension
 
 from tests.builders import MatchMSDataBuilder, PandasDataBuilder
 
 here = os.path.abspath(os.path.dirname(__file__))
 testdata_dir = os.path.join(here, 'data')
+
+
+@contextmanager
+def does_not_raise():
+    yield
+
+
+def _add_possible_key(keys, new_keys):
+    possible_keys = list(keys)
+    possible_keys += new_keys
+    return set(possible_keys)
+
+
+@pytest.mark.parametrize("filename, filetype, builder, expectation", [
+    ["Alkanes_20210325.msp", "msp", MatchMSDataBuilder, does_not_raise()],
+    ["Alkanes_20210325.msp", "csv", MatchMSDataBuilder, pytest.raises(NotImplementedError)],
+    ["Alkanes_20210325.dat", "msp", MatchMSDataBuilder, does_not_raise()],
+    ["Alkanes_20210325.csv", "msp", MatchMSDataBuilder, pytest.raises(Exception)],
+    ["Alkanes_20210325.csv", "csv", PandasDataBuilder, does_not_raise()],
+    ["Alkanes_20210325.csv", "msp", PandasDataBuilder, pytest.raises(NotImplementedError)],
+    ["aplcms_aligned_peaks.dat", "csv", PandasDataBuilder, does_not_raise()],
+    ["Alkanes_20210325.msp", "csv", PandasDataBuilder, pytest.raises(Exception)]
+])
+def test_filetype(filename, filetype, builder, expectation):
+    extension = get_extension(filename)
+    filepath = os.path.join(here, "data", extension, filename)
+    builder = builder().with_filename(filepath).with_filetype(filetype)
+
+    with expectation:
+        builder.build()
+
+
+@pytest.mark.parametrize("keys", [["test-key"], ["first_key", "second_key"]])
+def test_add_possible_keys(keys):
+    expected_ri_keys = _add_possible_key(Data.get_possible_ri_keys(), keys)
+    expected_rt_keys = _add_possible_key(Data.get_possible_rt_keys(), keys)
+
+    Data.add_possible_ri_keys(keys)
+    Data.add_possible_rt_keys(keys)
+
+    assert Data.get_possible_ri_keys() == expected_ri_keys
+    assert Data.get_possible_rt_keys() == expected_rt_keys
 
 
 @pytest.mark.parametrize("builder, filename_dat, filename_native, rt_unit", [
